+++ ipa_canopen_master/include/ipa_canopen_master/canopen.h
@@ -1,8 +1,8 @@
 #ifndef H_IPA_CANOPEN
 #define H_IPA_CANOPEN
 
-#include <ipa_can_interface/interface.h>
-#include <ipa_can_interface/dispatcher.h>
+#include <socketcan_interface/interface.h>
+#include <socketcan_interface/dispatcher.h>
 #include "exceptions.h"
 #include "layer.h"
 #include "objdict.h"
