+++ socketcan_interface/include/socketcan_interface/FastDelegate.h
