+++ socketcan_interface/include/socketcan_interface/socketcan.h
@@ -1,7 +1,7 @@
 #ifndef H_IPA_SOCKETCAN_DRIVER
 #define H_IPA_SOCKETCAN_DRIVER
 
-#include <ipa_can_interface/asio_base.h>
+#include <socketcan_interface/asio_base.h>
 #include <boost/bind.hpp>
 
 #include <sys/types.h>
